#62: Remove widget and undo removing widget now functional.
This was made easier through the new apis. I just needed to pass
the active contexts in somehow, which I did through query params. Once
I had them, it was just a matter of calling a few apis.

The hard part was getting the block to render properly when undoing
the command. For some reason, the same mechanism that was working
with Edit Widgets wasn't working here. It took some digging to find
the right sequence of functions to make it happen. Now we just need
some styling on the Undo block so it stands out.