#78: alter admin paths so node/edit and node/delete work in spaces
This is an issue with overlay in general and spaces. When Drupal checks a link to
see if it should be opened in an overlay, it doesn't know to remove the purl modifier
first. It wants node/5/edit and gets site01/node/5/edit instead. So it opens the link
normally.

By adding a wildcard to the front of each admin path, it should detect links in a space
correctly.