Applied previous fix to a few more classes #63
Didn't notice this was a problem before since we didn't have vsites to
worry about.  Announcement, class, and gallery are all fine now.  Book
is cuasing me to run out of memory (it wasn't using abstract preparerow
at all before.  Now that it is, it's running through inline images and
that's causing the error).  Working on that now.  Then checking biblio.

(Note my editor is having some trouble with producing commits that don't
get funky line breaks.  I'm working on it between commits.)