There were two issues here, the variable name for the default role is .._node_vsite instead of .._vsite_node now for default og roles.
The second issues is that the system module added a pre_init call to check and see if users have access to the administration theme, this cached the avalible roles for a user before vsite_init was able to run.  I added a call to reset the static avalible roles when they are modified. This should correct the permissions issues present in. #90