Adding the new module for the Harvard Activity Reports
This module allows faculty to import their publications contained in their OpenScholar site to whatever applications they chose via the XML.

At the moment this module has a few dependencies that are not yet functional in 7, so it cannot yet be enabled.  I was able to load those dependencies and needed tables in my local development to ensure the module is working as it should in 7.