site/register sets private site correctly.  #120
I was using the space's vsite_private variable to indicate privacy.  That variable is no longer in use.  Setting it will render the site inaccessible.  @rbran100 is updating the code that still used it that I'd been referencing.

The proper way to update privacy is by modifying the vsite node's group_access.  1 is private, 0 is public.