#132: Modify boxes' access functions to use new permissions.
It's just the one function that needs modification. We replace most of their interface for dealing with boxes anyway. There may be a few other places in boxes I'll need to modify to use the new permissions, but that should only be necessary if we send these patches to boxes.