#132: Replace various access controls we had with boxes_access.
I just needed to replace the access callbacks on our menu entries. I added a new menu entry for creation because its now a separate permission and needed to be treated as such. Luckily, I had designed cp_widgets to check for the truth of a can_edit or can_delete variable, which I can just fill with calls to boxes_access as appropriate. I also needed to replace our old means of determining whether a box can be created with checks against the permission.