#133: Add region permissions. Ensure Inline Widget Management respects permissions.
I did a number of things to make this work. For starters, I wrote a function to take a box and get its region. With that, I wrote an access function to check user permissions for the region, and placed it on the base path for the contextual links. In order for a contextual link to render, the base path must be accessible. This way, I was able to block all contextual links from showing with a single change to the menu tree.