Began porting os_simpleview #134
Ctools doesn't handle plugins the same as in d6.  I had to rewrite all the hooks advertising that this module had its own type of plugin.  I also couldn't find support for class based plugins loading parent classes (!!) so I had to add that too.

Rewrote a couple helper functions my plugins were making use of.  One was a wrapper for os_get_bundles.  The other is still in progress.  It gets the menu link for a node type within a vsite.  Only we're not within a vsite anymore, so it needs a little more re-thinking.

Forms load from the plugins just fine.

Currently stuck in the multiform itself.  It gets stuck in a loop of trying to redirect to the next page.  I can detect and fix this when the page advances, but that keeps it from working when going back a step.  I remember this being problematic in 6 too.

Haven't looked at the box or view yet.  I'd like to have the form done first.