Fixed simpleview wizard redirection loop #134
Set a flag in the form state cache to note when the form is advancing versus when it has advanced.  When it's already advanced, prevent redirection to next page.  Next and back work now.

Next: form completion needs to save box.  Box and ctools ajax calls need updating.  Also wondering if there's a good way to save changes to an old box and create a new one with similar code.  Also need to make sure the box itself works.