#131: Dismiss the feature customize form when its submitted.
Turns out this is an API change that happened with ctools in 7. Previously, its form wrapper would return an empty array. We could determine if the form submitted successfully by whether this array was empty or not. Now it always returns the array, and we need to check form_state['submitted'] to determine if the form submission was successful. This is the first modal I've had to deal with, since so much of our interface uses overlay now.