Fixed last ajax alert #130
Cleaned up that one last ajax problem.  Here's how I was duplicating it in firefox and chrome.

1. Log out.
2. Go to site/register
3. Enter a known bad domain name.  You should see a domain invalid error after leaving that field.
4. Use good values for name, email, password.
5. Submit the form.  Get reminded about domain.
6. Enter a new domain name.
7. Without clicking anywhere else (as that will trigger domain events), click submit.
8. When submit returns, immediately click the link to go to your site.
9. Try to read the alert really quickly before the next page loads.

The fix in the thread worked, without patching core or overwriting window.alert.  It changes the settings for what to do on an ajax error.  Instead of a alert, it does nothing now.