#60: Changed my mind.
Quicktabs is a big module for some reason. It has a lot of classes, a lot of javascript, and a lot of code that we just don't need. It's changed a lot since 6, and worst of all, it uses its own custom tabs script. Which is dumb because jquery.ui.tabs is in core now. I'll just render everything myself and use jquery ui tabs to do the switching.