Began theme migration #150
Copying og.og_theme to spaces_overrides.  How hard can it be?

Turns out it's uglier than expected.  I started with the Table destination.  The problem with that is that it requires a primary key, which spaces_overrides lacks.  Altering the schema of spaces seemed out of bounds for migrate, so...

I used TableCopy instead.  TableCopy requires me to provide a schema, but doesn't depend on a primary key.  I couldn't map default values, but had to alter the row during prepare instead.

The other part that was tricky was that TableCopy blindly writes whatever properties its object has to the db.  The map object keeps a record of the primary key of source data (which it uses for rolling back migrated data).  The problem here was that I had different columns.  og_theme is not the same as spaces_overrides' value.  To get around this I aliased the two fields I was getting, so that as far as migrate is concerned, we're only ever dealing with id and value, even though they're coming from nid and og_theme.

Dropped the scholar_ prefix with a regex.  Filtered out themes that aren't enabled in the current installation.  They get ballard instead (although I'd love to have an OS_THEME_DEFAULT or something).

Unsupported themes throw a notice.  The three dropped themes did this correctly.  Also got notices from agordon, theme_01, leap, and rbates.  Unsurprising since they're one offs.

Todo: flavors.  Will probably do this as a separate migration class in the same file.  I'll have the same spaces_overrides problem and would like to keep classes that use that hack as small as possible.