#60: Switching tabs works correctly.
Always something simple. I was having trouble because switching tabs would cause the entire page to be reloaded in the tab. Turns out it was because Drupal rendered more of the link href than I wanted it to. UI Tabs thought the link was a remote link, and loaded it in via ajax. By setting the EXTERNAL option to TRUE on the tab links, I was able to force Drupal to print what I wanted and nothing else, and that fixed the ajax loading behavior. All that's left is to make sure the sessionStorage thing still works and then I can ship it off to QA.