#60: Tabs remember which tab they were on.
Made easier since we're now using jquery ui tabs. I couldn't set a handler on the select event, because it happens too early and I couldn't get the current id from it. Instead I set an event on the show event to set sessionStorage. And then when we initiate the tabs on page load, I could just read from sessionStorage and set it as the default.