#93: Ensure shields directory is created before trying to move files into it.
We never created the directory, so when Drupal went to find the real path to the destination, it had trouble since the path requested didn't exist. As a result, it crapped out. So now when we save shields, it attempts to create the directory they go into first. If this fails, there's some deeper issue with the server and we notify the admin of this.