#156: Change boxes admin paths so they activate cp theme.
I didn't want to change the paths. But the function that sets the cp_theme only does so based on path. I expected it to be based on context, but I guess the context hasn't been set when hook_custom_theme is run. So, I changed all the paths from os/boxes to cp/boxes.