Migrated updates context names #144
Continued updating context migration in spaces_overrides.  This involved two things.

First of all, we don't store contexts the same as we used.  In d6, they had context:reaction:block:theme as the key.  We've since dropped the theme part.  What this means is that a single vsite might have had a different context for each theme, all but one of which should not be migrated.  So I added a check to see what theme was in use in the previous vsite.

The only other factor was the context name itself.  I dropped the openscholar-feature- part and repeated the name of the feature, as per current convention.  There's also a map of contexts whose names don't apply to that (ie os_public).

Rewrote the theme name update as a function before realizing it wouldn't apply to contexts.  Seems like it might be needed later anyway.

According to Matt, the context only needs to provide a block's 'module', 'delta', 'region', 'weight', 'status', and 'title' (and the title should be '' so it can be over loaded).  So I filtered out the rest.  The region is updated as before.  The module name is updated as well, but I'm sure there are some exceptions where this doesn't work, like if the module no longer exists or was renamed oddly.