#132: Use boxes_access when getting the factory list.
I was calling user_access to get the access level for the factory list, since all I had was the plugin_key, not an actual box. This missed the catch-all 'administer boxes' permission, though. I made a dummy box and ran that through boxes_access and now the catch-all works.