Moved the check for active theme to query #144
I was checking that a context reaction used a vsite's current theme during prepareRow, after the query took place.  Moved it to the query itself, and cut out a third of the contexts to migrate.  (Also sped things up by removing the legacy db query that took place for each row...)

There was some weirdness with matching a context id that ended with the og_theme.  Long story short I had to use $query->where() instead of $query->condition as well as a placeholder.  See comment in commit.

I misinterpreted vsite-section-public.  We're not using that in overrides.  openscholar-feature-none is the d6 context to use for os_public (aka widget appears in all sections of site)

Looks like the only thing left to do in context is updating the box deltas.  I'm not even sure if that's something that needs to happen, or if we can use the old deltas.  At any rate, that can happen in the _update_box function which I'll attend to when I get to boxes themselves.