Began migrating menus from spaces_overrides #144
Began menus.  Also created a boxes class, but it does nothing so far but query for boxes.

Ignoring the "Not in a menu" menu.  It is no longer in use so I'm not even querying for those menu items.  If we have extra time, I could make a third menu for those, but not add it to the layout.

Menu names were straightforward.  features -> primary-menu and scholar_secondary_menu -> secondary_menu.

Most of the data in each menu has stayed the same.  The big difference is that menu items have an mlid.  For custom links, this is forged by vsite_menu.  I forged it the same way.  I haven't gotten the mlids from inherited menu items yet.  Also ignored the depth and p2-p9 parent settings for now.