spaces_overrides gets mlids of existing menu items #144
First stab at spaces_overrides menu migration forged all the mlids.  Now I'm grabbing existing ones from the db.  I tried to use os's menu plugin, but was working within a vsite and I needed to write directly to the db instead.

cp/menus doesn't actually work without a preset.  Updated variable migration to adjust the preset names.  Now some of my menu items are loading.  I'm going to pick a vsite or two and focus on getting their menus to migrate perfectly (or as close to it as I can for now.  I don't have a menu item in the system for profiles yet, since we don't AFAIK have that feature.  So I can't use the system's link there.)

Fixed up a couple of my helper functions along the way.

TODO: vsite_menus uses a spaces variable to store the last used mlid so it can be incremented the next time we need to fabricate one.  I'm storing it the same, but it never gets cleared after a migration is rolled back.  This isn't urgent, but I should reset it after a menu is removed from the overrides table.