verify initial batch release imports all and cron new release imports
still a little work to do to get it to redirect properly on initial
import.

idea: what if the batch process took care of the initial release AND
the cron job too? this would also fix the scenario in which two
releases happen to fall between two cron runs -- which would leave an
orphan package on rebuild without importing. #155