Began boxes #144
Began updating boxes.  I haven't done any mappings yet, but check that a box's plugin key is in os_boxes_get_boxes_plugins().  Warnings when it's not.  I plan on going through the warnings and writing mapping for each of those (or actually migrating into a different box type when that doesn't work).

Found a couple error cases:

There are boxes with no object_id and very little box info.  By very little, I mean they have an array of node titles, nids, and weights.  I think they used to be featured posts widgets.  Maybe there weren't completely deleted?  I'm ignoring them.

Also found a few rows that couldn't unserialize.  That's because the row was cut off.  This happened to single images where the image was included inline as data.  MySQL text fields are apparently limited to 64kb, as that's where the image is truncated.  I drop these as well with a notice.