#166: Add advanced settings permission. Use it to control access to a fieldset.
I defined a permission, and a fieldset in os_boxes_default to use this permission to control access. Developers can put additional fields in there that would be too confusing or high-leveled for normal users. All they have to do is call the parent's options_form function and add the fields to form[advanced].