Made a few changes to the code because I realized that the code was never looking for the global id, the code was only looking for the local users GA account.  I'm not sure why I didnt catch this earlier, but it has been caught and changed.  The code now properly checks for a global GA account (vsite_ga_global_id) and a local GA account (vsite_ga_google_analytics_id).  It properly displays the code if one or both is found.  For some reason the admin form for the GA module was also setting the local variable (vsite_ga_google_analytics_id) rather than the global id, so the forms were overwriting one another, this has since been corrected.
The only buggy thing I've found is in the local cp_setting version of setting the GA account, the user has to click "Save for OS Personal" in order for the GA account to properly show up, I'm not sure if this is the correct method or not and will resolve this first thing in the AM 8/24.  So I'm switching the status of this back to "in dev" until I can confirm the proper process.  #123