spaces_overrides:boxes gained support for porting node boxes. #144
Added conversion for os_boxes_node to simple boxes.  Tested locally and handled all the failing edge cases (nid=0, nid => not a node, nid => vsite).

I'm doing something similar to what I did with variables.  My class defines _port_box_$PLUGINKEY functions.  If one of these exists, the box is passed off to that function.  Any changes to the box structure should be done in one of these functions.  If a _port_box function does not exist, the plugin_key is checked against a list of mappings and against the list of enabled plugins.  If no plugin key map is found, we get an error about that box type not being supported.  The plan is to add porting functions and key mappings until all those errors are gone.