#172: Modify nice menus settings form to support new dropdown var.
I figured its only relevant when Nice Menus is enabled, so it was safe to put the form in Nice Menus' admin form. The form works correctly, and it disables the relevant functionality when disabled, so that's good. I'm not sure what to do with children links on the menu form though. Should they be flattened, should they be hidden, or something else entirely?