Migrate ports os_boxes_modal_sitelogo #191
I already copied the site logo during vsite migration, so this wasn't too hard.  I did take the shortcut of loading a vsite node instead of querying the db for all the data I needed.

I had 6 instances of vsites without logos.  Obviously these didn't work.  It would be nice to programatically determine if the logo failed to copy during vsite migration or if the site simply has no logo.  But for 6 instances of an edge case it's not worth the trouble.