#193: Site Info reads from vsite fields, never saves box while in vsite.
There was no way I could tell boxes to not save a box. I had to circumvent it entirely, using the boxes_form method. Boxes_box_form will call this function, then immediately return the output. It comes with none of the window dressing normally applied to these boxes. This meant I had to provide my own buttons and submit function, which is fine. It's actually cleaner now, since the user won't ever be able to change the title or description, both of which would create entries in spaces_override.

Now on to site logo. It was trickier the first time around. Hopefully it won't be as tricky this time.