Stopped unsetting region in boxes migration #144
Region -> false was necessary to keep an unused box out of a region.  I thought having it in context would be good enough, but it needs to be in the box too.

Latest announcements keeps getting imported.  It exists in the original site, but the announcements feature is disabled.  It's disabled in the new site too, but this isn't preventing the box from showing.  Maybe it's an sv_list thing (sv_list hasn't been ported yet)