Fixed manual list (display option).  Began shield port.  #144
The manual list box was throwing notices over options['style'] not existing.  So I added it during migration.

Began porting shields.  I'm assuming we'll update shields through the shield manager before migrating.  I could do a shield migrator before then, but that seems like it's low priority.

Here's the problem with shields.  The selected shield lives in a variable in d6.  This means that once again, there's no override to migrate into the new box (which stores the selected shield as an option).