OS site logo uses original instead of pre-cropped image.  #191
Site logo was getting cropped wrong because it was using the thumbnail from d6 instead of the full image.  I've switched to the full image (which wasn't saved in db anywhere but had .crop.jpg or something appended to the filename) and migrate that instead.

That copies fine, but doesn't create a cropped image.  So now the full image shows up and if you edit the box you see the intended cropping.  Looks like I'll have to create that after copying the image.