#162: Extra links now show up in Firefox.
I had no idea the drawer links weren't showing up in Firefox. So I reworked it to the way I had originally intended to do it, using the top property instead of margin-bottom, and fixed the bug that was causing it to fail in Chrome. It looks great in both.