sv_list_box has a working pager #224
EntityFieldQuery is smart enough to work with the standard drupal pager.  When $EFQ->pager($num) is called, further calls to theme_pager will output the efq's pager.

To get this using ajax didn't take too much more code.  I had to give the pager links the use-ajax class and send them to my sv_list ajax callback.

The ajax callback will, in the absence of javascript advance you to the page the dfeault pager would have.  With javascript it'll load the sv list box and render it.  Conveniently, theme_pager relies on the ?page argument, so that's already in place.  ajax_command_replace and we're done.

The only downside to this approach is that theme_pager didn't let me pass options to the links it generated.  I did it with a regex instead.  I'd like to look for other options before finishing this one.