Post in widget does drupal_render on migrated node #234
Post in a widget now requires node migration to be complete before it can be run.  This way we can just do a local drupal_render and save the results to an html box instead of querying the legacy database.

I kept the legacy query as a backup if node load fails.  I think this might let us get around porting bio nodes.  Just treat the bio PIB widget as a PIB and then the legacy query will get the bio text and save to an html widget.  This only makes sense if we're making bio a widget and not converting it to a page.  Still not sure if there's a decision there.