More link text comes from ajax, not js.  #224
Made an ajax callback to get the more link text given a content type.  It's hardcoded for now.  We should have an API function for this at some point.

Also removed the ajax callback for getting a rendered box after paging.  That will have to be implemented differently and the nonfunctional code would have been misleading.