Content type/more link ajax preserves visibility #224
When I changed the content type, the more link textfield was replaced.  This included replacing the display:none that form states had applied.  Form states weren't checked following the ajax call.

I couldn't get the form states to retrigger, so instead of replacing the entire form element, I just replaced the input.  Display:none was in the input's parent, so now the input can be swapped in and out regardless of the state of the show more link checkbox.