Gave sv list a filter for documents.  fixes #249
I couldn't filter in the EntityFieldQuery because menu depth is not a field.  Created a separate query to get a list of book pages with menu depth 1.  Added the results as a condition on the EntityFieldQuery.

Doesn't quite work yet because of #250.  Works great if I remove the spaces condition, so I'm closing this ticket.