#246: Ajaxify sv list pager.
EFQ took care of the paging part on its own. I just needed to rerender it and return the result. I opted to not use Drupal's ajax mechanisms because they're designed to work from a render array. I couldn't insert the ajax stuff into the render array, and making it work otherwise would be a lot of trouble. It was easier to just circumvent it and use jQuery's ajax handling instead. This way, I could be sure that url got the right parameters, that the client-side caching worked properly, that nothing interfered that I'd have to deal with.

All that's left is to figure out a throbber when the ajax is in progress.