site/register clears leftover errors after success #287
site/register was letting drupal's password confirm field validate itself.  This meant it was setting errors the non-ajax way even though the page was delivered via ajax.  The consequence is that on the next page load, the errors showed up.

To fix this I cleared errors after the form succeeded.