Migrated cv #301
CV attached files were irritating.  There were two fields - attach and attached.  In most cases they were identical, but there were 34 nodes that had differing files.

I checked prod and it looks like attach was the one that was in use.  I migrated that one, but added a condition that if the file couldn't be found, use the other one instead.