Ctools modal are given position: fixed.
Modals should always be on screen. When the page is long, like with an edit form, the modal would be at the top of the page. So you'd click 'Edit File', the screen would gray out and you'd see nothing. By making the modal position: fixed, the modal will always be on screen.