Clean up delete node form.
Removed the local tasks. This was something overlay does for core, but adaptivetheme is wierd about it so puts them back in.
Removed the tabs on top. Ugly and useless. No one's getting around the overlay system through them.
Removed the duplicated title. Unnecessary with the title hanging above the overlay proper.
Change cp theme function to check for node/delete as well as node/edit.