Began software project, release #301
Project was fine.  Encountered something new in release.  The fid mentioned by a release can point to an entry in the files table with a filepath on a remote server.  I didn't realize that was an option, but migrate does not approve.  There are over 1k of these so I have to figure out what to do with them without resorting to wget whenever a filepath starts with http.