#52: Media embed widget functional.
Media provides a form array element, called 'media'. It functions exactly the same as the file field widget, except for the multi value thing. We don't want that anyway.

It still opens a popup within a popup, but there's really no way around that without a lot of work for a very edge case.

On the plus side, this widget can pull double-duty for Media Embed AND Single Image.