#228: Make the spaces current space filter work with files.
This required two things, the latter of which took me forever to find. The first was to alter the view data so spaces could be joined with the file_managed table. The second was to alter the pre_view hook in spaces_og to allow for the membership being swapped when the base table was file_managed. I had trouble just finding the pre_view hook. Once I found it, the filter works great.