Box removal path needed updating.  #286
AddThis used the path to determine if it should issue a page reload.  The box path changed in #318, so reloads were no long being sent when the box was redisplayed.

The reload was also happening in the modal, which resulted in the wrong location being loaded.  Reloaded the parent window instead.