slideshow form fix.  #163
The os_boxes_slideshow form uses a function from views_slideshow_cycle to get the list of effects.  When Dan tried to create a slideshow, the include file with that function was not available, so the function didn't exist.  Added some checks to include the file if needed and to avoid the WSOD.  Didn't have this problem on my own installation though.