#30: Fix contributor widget 'Add Another Field'
For some reason, the js attached to this stopped working. I set the value of a hidden field with it, and without that value, the form fails validation, and ajax returned nothing. I still don't know why the js stopped working, I rewrote it to use a different event that hopefully won't be broken by anything in the future.