Fixed recent docs sql condition.  #369
SV list has an exception for books.  It does some extra sql do get a list of top level book nodes so it can filter out the pages.  When no top level nodes exist, it was trying to add an impossible condition, hence the error.  Only applies that condition when book nodes with depth 1 exist.