Migrated comments #344
Comments are migrated.  This was almost straightforward.  Comment migration requries a bundle.  This meant doing a dynamic migration, with each node type registering a comment migration instance.  I had that happen in each node's post import function and only if comments existed for that node type.  The only other problem was that the bundle had to be comment_node_$type.  I was just passing it the $type.  This meant that comment default fields were available for migrating, but there was no body field for them since they weren't a real comment bundle.

I have not done the comment_sources variables portion of the ticket.