Fixed slideshow pager numbers #163
Fixed the pager text.  Disagree with removing the counter.  It doesn't make sense with the pager, but the pager is not mandatory.  Some people might prefer next/back with a current page status.

Also fixed the gallery nid.  It was setting an argument, but the view didn't take one.  Not sure what hcanged between now and when we last tested it.