Began adding system variable migration #382
Added Variables migration class to tables migration.  Browsed through list of variables and grabbed likely candidates.

Updated tablecopy with an option to ignore rows if a class's depenencies aren't enabled.  Would like to rework this into a per class thing instead of per row.

OSMigration parent class gained a d6_list_content_types() method.  Surprised this is the first time I needed it.  There is a comments variable for each node type, so this was necessary when filling out the whitelist of variables to migrate.

Updated shields to work with Richard's latest change, which affects the table alias.