Migrated usage log #385
Migrated usage log.  Used my new module dependencies option.

Only problem was that a class without the dependency still queries the old db to get a count of how mnay rows are importable.  In this case that was a problem since the old db doesn't exist on my prod backup (it only existed in projects).  To keep the error from being displayed I made a small hack.  When the legacy table doesn't exist, it queries node for a null nid.