Migrated vsite's field_group_path.  #406
Added group path.  First we check for a vsite_domain in purl.  If that's not found we try the vsite_domain_name override variable, and prepend that to the original purl (this case being a group domain).  If neither of those is available, we prepend the purl_base_domain to the original purl value for a standard vsite with no domain shenanigans.

@rbran100, can you check this over?  Not a lot of room for error on this one.