#373: Load public css and flavor into wysiwyg.
Used a recursive method to get the theme css and the css from its ancestors. Then add the flavor css in it. The get_flavor function only works on the currently active theme, which will be in cp in all cases, so not helpful. I had to get it manually.