TableCopy restricts query for missing module deps.  fixes #414
Previous version was trying too hard to be clever.  When a required module didn't exist, it was invalidating the query by getting all rows from the table where the primary key was null.  This didn't work since the table wasn't likely to exist in this case.  Instead I'm now searching node where nid is 0 to get 0 rows.

Tested this out with a fake migration class for a table that doesn't exist.