Fixed path of inline files for migration.  #417
In migrate 2.2, inline files had to be migrated over the internet.  It wouldn't do local file copy (or if it did, I never figured out how).  Migrate 2.4 fixes this.  All I had to do to fix the problem was swap out the purl_base_domain for the path to the files.

Tested locally.  Still needs HWPI testing, since that's where the problem was discovered.