Fix bug with shields on reinstall.
The bug only occurs when reinstalling OS where an old install that had registered shields was. The admin form read the shields on the hard drive and used those to generate its list. Drupal proper had no idea they were there, so there was no data in the variable to go with them. When you save the form with these phantom shields, the fid would be null. When we go to render the shields, the fid is still null, and Drupal throws errors when this is the case.

For such a long confusing description, the fix was easy. Generate the list from the variable instead of the hd.