Profile migration uses known locatiosn of thumnbnails for big images.  #340
Made the thumbnail finding function a little more robust.  Provides notice for images that were swapped to a thumnbnail and warnings for images that had no thumbnail and were abandoned.

After that I made a list of images recieving warnings and locations to their thumbnails.  Short of using a different image toolkit I think this is the best way to do this.