#473: Fix pager with certain boxes
The issue is that the underscores in the box's delta were converted to hyphens by Drupal. There's no box like that, so it crapped out. I changed the way the script gets the delta to something more robust, and the problem is resolved. Also, I fixed an error that could happen if the user clicked on the numbers. It didn't affect operation in any way, it just spat out an error.