#473: Bug fixes with pager.
There are two issues fixed here.
The first: if there were any extra ajax commands, the pager would fail. I expected to always use the 2nd command given, the first being the settings array. On staging, a different command is second, causing failures. This was fixed by checking every command for the one we want.

The second issue is that the pager would load the first page again if there was another pager on the page. The pager system assumes we're reloading the whole page, so it includes the pages for every pager on the page. If there are two pagers, [page] would be num,num. The box didn't know what to do with this, and defaulted to the first page. After this first page-swap, it would be fine. To fix it, I passed the id of the box's pager to the js, get the page for our pager, then pass only that to the our callback.