Citation distribute.  Upgrade dash plugin.  #395
That was worse than expected.  The path from a node being saved to distributed goes something like this:

citation_distribute_node_insert() flags the node
citation_distribute_node_flag responds to the flagging, calling functions to push content.

So far so good, except that flagging a node takes a nid as a parameter.  By the time the final pushing functions are called, we have a nid, but no node.  Any data from the node that needs to be pushed has to be acquired with a node_load.

There were two problems with this that weren't present in d6.  field_upload wasn't available yet.  It was present during the insert hook, but node_load wouldn't retrieve it at this point.  DASH requires a full text attachment, so we need to see files.

I went with the quick and dirty method to get around that.  When we flag the node, a static copy of the node is saved.  Later on, if a field we need is absent, we check the static copy.

The second problem is that file entities weren't actually created yet.  The above method gave me a fid, but file_load returned false.  However, some of the file data was already in file_managed.  Again opting for the quick and dirty method, db_select got us the uri and mimetype.

I'm not in lover with either of these compromises, but I still don't have a better idea.  I think removing flag would fix this problem (and alleviate some of the complexity of making sure flag and CD's table are in sync) but that's nontrivial and would have to wait till we have more time.