Migrate creates media for os_boxes_media #424
Spoke too soon.  Still awful.

I added support for embeds that were just a url instead of text.  I figured media would handle giving me the embed code for them and then I could save it same as before.

Nope!  When I save the MediaInternetOEmbedHandler, if the url is not valid (ie if it's a youtube video that's been taken down or an openscholar url that wasn't embeddable to begin with), the file is still saved.  Sometimes I get an exception, sometimes not.  Oh and the saved file has no type, so viewing admin/content/files or attempting a file_delete failed.

The solution was to use a db_transaction that I could rollback when an exception was thrown or a file saved with no type.  That part worked as advertised.

Also wrote rollback functions to clear out my custom files between migrations.