#488: Create new view modes for files that are more understandable.
File entity has a somewhat convulted system for this. First you define a view mode, then a formatter, then a 'file display' which is the link between them. A file display is really just a string composed of {file_type}__{view_mode}__{formatter}. If the display matches the file_type and view_mode arguments, then the formatter gets run. If that formatter doesn't output anything, then it falls through to the next display in the list that matches the arguments and tries that formatter.