Fixed migrate's imagefield crop handler.  #464
I think I've got the crop handler working correctly now.

First we check for the presence of an original image.  If we have it, crop it.  I was providing default dimensions, but those proved to be incorrect.  We still need to pass dimensions along, so I'm using the full size of the image and letting the preset scale it down.  If all that is successful, we set the file class to be fid based.  Migrate provides several file classes for the myriad ways you can migrate files.  The default is by URI.  In this case, the file is already migrated, so we just need to supply the fid.

If there is no original, we use the thumbnail provided, but don't crop it.