#513: HTML files use large file icon formatter for preview view mode
If a file view modes doesn't define which formatters to try, they use the file_link theme. This breaks the Media browser, on top of looking ugly. I added a display to use a different formatter, and it rendered using that. Media Browser selection now works and it looks much better as well.

On to the view mode select form.